Implemented a Centralized Main and SceneManager
Now, no other file in the project has a main method (except for JDBC which we have not yet integrated).

- New FXML/SceneBuilder scenes should have only a controller class and the fxml that comes with them

- New Pure Java scenes should have
  only a basic concrete class with a static method called getRootNode() for the scene manager to call and implement the scene in its switchTo() method.

ANY TIME a new scene is made, PLEASE add an id to the SceneID enum file and create a case block in SceneManager's switchTo() so it knows what to do with that scene when it is switched to.

Thank you :)