Hooked Up General Screen and Login
I created a very basic General screen with a table view that shows all the events. (Currently commented out the WHERE clause that limits it to the active ones for testing)

I gave it a scene id and hooked it up with the scene manager, so now when you successfully log in, you're brought to the general screen.

I also started (but commented out) some code that will keep a static User activeUser that can be used to keep track of who is signed in.